Update from original

+++ run_tests.py
@@ -59,10 +59,15 @@ def run_command(cmd):
     lexer.whitespace_split = True
     lexer.escape = ''
     arg_list = list(lexer)
-
-    sp = subprocess.Popen(arg_list, stdin=None,
+    
+    try:
+        sp = subprocess.Popen(arg_list, stdin=None,
                           stdout=subprocess.PIPE,
                           stderr=subprocess.PIPE)
+    except: 
+        print_debug("ERROR: The child (%s) raised an exception: %s\n" % (cmd, sys.exc_info()[1]), s, run_tests_log)
+        raise
+
     out = sp.communicate()
     output = ""
     output += out[0].decode("utf-8")
@@ -80,10 +85,11 @@ def run_cmds(compile_cmds, run_cmd, filename, expect_failure):
             if output != "":
                 print_debug("%s" % output.encode("utf-8"), s, run_tests_log)
             return (1, 0)
-
-    (return_code, output) = run_command(run_cmd)
-    run_failed = (return_code != 0)
-
+    if not options.save_bin:
+        (return_code, output) = run_command(run_cmd)
+        run_failed = (return_code != 0)
+    else:
+        run_failed = 0
     surprise = ((expect_failure and not run_failed) or
                 (not expect_failure and run_failed))
     if surprise == True:
@@ -156,8 +162,12 @@ def run_test(testname):
     # is this a test to make sure an error is issued?
     want_error = (filename.find("tests_errors") != -1)
     if want_error == True:
-        ispc_cmd = ispc_exe_rel + " --werror --nowrap %s --arch=%s --target=%s" % \
-            (filename, options.arch, options.target)
+        if (options.target == "knc"):
+            ispc_cmd = ispc_exe_rel + " --werror --nowrap %s --arch=%s --target=%s" % \
+                (filename, options.arch, "generic-16")
+        else:
+            ispc_cmd = ispc_exe_rel + " --werror --nowrap %s --arch=%s --target=%s" % \
+                (filename, options.arch, options.target)
         (return_code, output) = run_command(ispc_cmd)
         got_error = (return_code != 0)
 
@@ -211,7 +221,7 @@ def run_test(testname):
                     obj_name = "%s.obj" % os.path.basename(filename)
                 exe_name = "%s.exe" % os.path.basename(filename)
 
-                cc_cmd = "%s /I. /I../winstuff /Zi /nologo /DTEST_SIG=%d %s %s /Fe%s" % \
+                cc_cmd = "%s /I. /Zi /nologo /DTEST_SIG=%d %s %s /Fe%s" % \
                          (options.compiler_exe, match, add_prefix("test_static.cpp"), obj_name, exe_name)
                 if should_fail:
                     cc_cmd += " /DEXPECT_FAILURE"
@@ -242,20 +252,28 @@ def run_test(testname):
                         and (options.include_file.find("knc-i1x16.h")!=-1 or options.include_file.find("knc.h")!=-1 or options.include_file.find("knc2x.h")!=-1):
                     gcc_isa = '-mmic'
 
-                cc_cmd = "%s -O2 -I. %s %s test_static.cpp -DTEST_SIG=%d %s -o %s" % \
-                         (options.compiler_exe, gcc_arch, gcc_isa, match, obj_name, exe_name)
+                if (options.target == "knc"):
+                    cc_cmd = "%s -O2 -I. %s %s test_static.cpp -DTEST_SIG=%d %s -o %s" % \
+                         (options.compiler_exe, gcc_arch, "-mmic", match, obj_name, exe_name)
+                else:
+                    cc_cmd = "%s -O2 -I. %s %s test_static.cpp -DTEST_SIG=%d %s -o %s" % \
+                         (options.compiler_exe, gcc_arch, gcc_isa, match, obj_name, exe_name)                    
+
                 if platform.system() == 'Darwin':
                     cc_cmd += ' -Wl,-no_pie'
                 if should_fail:
                     cc_cmd += " -DEXPECT_FAILURE"
-
-            ispc_cmd = ispc_exe_rel + " --woff %s -o %s --arch=%s --target=%s" % \
-                       (filename, obj_name, options.arch, options.target)
+            if (options.target == "knc"):
+                ispc_cmd = ispc_exe_rel + " --woff %s -o %s --arch=%s --target=%s" % \
+                           (filename, obj_name, options.arch, "generic-16")
+            else:
+                ispc_cmd = ispc_exe_rel + " --woff %s -o %s --arch=%s --target=%s" % \
+                           (filename, obj_name, options.arch, options.target)
             if options.no_opt:
                 ispc_cmd += " -O0" 
             if is_generic_target:
                 ispc_cmd += " --emit-c++ --c++-include-file=%s" % add_prefix(options.include_file)
-
+             
         # compile the ispc code, make the executable, and run it...
         (compile_error, run_error) = run_cmds([ispc_cmd, cc_cmd], 
                                               options.wrapexe + " " + exe_name, \
@@ -263,13 +281,14 @@ def run_test(testname):
 
         # clean up after running the test
         try:
-            if not run_error:
-                os.unlink(exe_name)
-                if is_windows:
-                    basename = os.path.basename(filename)
-                    os.unlink("%s.pdb" % basename)
-                    os.unlink("%s.ilk" % basename)
-            os.unlink(obj_name)
+            if not options.save_bin:
+                if not run_error:
+                    os.unlink(exe_name)
+                    if is_windows:
+                        basename = os.path.basename(filename)
+                        os.unlink("%s.pdb" % basename)
+                        os.unlink("%s.ilk" % basename)
+                os.unlink(obj_name)
         except:
             None
 
@@ -295,11 +314,13 @@ def run_tasks_from_queue(queue, queue_ret, queue_skip, total_tests_arg, max_test
 
     if is_windows:
         tmpdir = "tmp%d" % os.getpid()
+        while os.access(tmpdir, os.F_OK):
+            tmpdir = "%sx" % tmpdir
         os.mkdir(tmpdir)
         os.chdir(tmpdir)
     else:
         olddir = ""
-        
+    
     compile_error_files = [ ]
     run_error_files = [ ]
     skip_files = [ ]
@@ -323,7 +344,12 @@ def run_tasks_from_queue(queue, queue_ret, queue_skip, total_tests_arg, max_test
             sys.exit(0)
 
         if check_test(filename):
-            (compile_error, run_error) = run_test(filename)
+            try:
+                (compile_error, run_error) = run_test(filename)
+            except:
+                print_debug("ERROR: run_test function raised an exception: %s\n" % (sys.exc_info()[1]), s, run_tests_log)
+                sys.exit(-1) # This is in case the child has unexpectedly died or some other exception happened
+            
             if compile_error != 0:
                 compile_error_files += [ filename ]
             if run_error != 0:
@@ -378,7 +404,7 @@ def file_check(compfails, runfails):
         compiler_version = options.compiler_exe + temp3.group()
     else:
         compiler_version = "cl"
-    possible_compilers = ["g++4.4", "g++4.7", "clang++3.3", "cl"]
+    possible_compilers = ["clang++3.3", "clang++3.4", "cl"]
     if not compiler_version in possible_compilers:
         error("\n**********\nWe don't have history of fails for compiler " +
                 compiler_version +
@@ -452,12 +478,12 @@ def verify():
     f_lines = f.readlines()
     f.close()
     check = [["g++", "clang++", "cl"],["-O0", "-O2"],["x86","x86-64"],
-             ["Linux","Windows","Mac"],["LLVM 3.1","LLVM 3.2","LLVM 3.3","LLVM head"],
+             ["Linux","Windows","Mac"],["LLVM 3.1","LLVM 3.2","LLVM 3.3","LLVM 3.4","LLVM trunk"],
              ["sse2-i32x4", "sse2-i32x8", "sse4-i32x4", "sse4-i32x8", "sse4-i16x8",
-              "sse4-i8x16", "avx1-i32x8", "avx1-i32x16", "avx1-i64x4", "avx1.1-i32x8",
+              "sse4-i8x16", "avx1-i32x4" "avx1-i32x8", "avx1-i32x16", "avx1-i64x4", "avx1.1-i32x8",
               "avx1.1-i32x16", "avx1.1-i64x4", "avx2-i32x8", "avx2-i32x16", "avx2-i64x4",
               "generic-1", "generic-4", "generic-8",
-              "generic-16", "generic-32", "generic-64"]]
+              "generic-16", "generic-32", "generic-64", "knc"]]
     for i in range (0,len(f_lines)):
         if f_lines[i][0] == "%":
             continue
@@ -529,8 +555,9 @@ def run_tests(options1, args, print_version):
     ispc_exe += " " + options.ispc_flags
 
     global is_generic_target 
-    is_generic_target = (options.target.find("generic-") != -1 and
-                     options.target != "generic-1" and options.target != "generic-x1")
+    is_generic_target = ((options.target.find("generic-") != -1 and
+                     options.target != "generic-1" and options.target != "generic-x1") or 
+                     options.target == "knc")
     if is_generic_target and options.include_file == None:
         if options.target == "generic-4" or options.target == "generic-x4":
             error("No generics #include specified; using examples/intrinsics/sse4.h\n", 2)
@@ -551,13 +578,18 @@ def run_tests(options1, args, print_version):
             error("No generics #include specified; using examples/intrinsics/generic-64.h\n", 2)
             options.include_file = "examples/intrinsics/generic-64.h"
             options.target = "generic-64"
+        elif options.target == "knc":
+            error("No knc #include specified; using examples/intrinsics/knc-i1x16.h\n", 2)
+            options.include_file = "examples/intrinsics/knc-i1x16.h"
  
     if options.compiler_exe == None:
-        if is_windows:
+        if (options.target == "knc"): 
+            options.compiler_exe = "icpc"
+        elif is_windows:
             options.compiler_exe = "cl.exe"
         else:
             options.compiler_exe = "clang++"
- 
+
     # checks the required compiler otherwise prints an error message
     PATH_dir = string.split(os.getenv("PATH"), os.pathsep) 
     compiler_exists = False
@@ -575,7 +607,30 @@ def run_tests(options1, args, print_version):
         common.print_version(ispc_exe, "", options.compiler_exe, False, run_tests_log, is_windows)
  
     ispc_root = "."
-    
+
+    # checks the required environment otherwise prints an error message
+    if (options.target == "knc"):
+        options.wrapexe = "micnativeloadex"
+        PATH_dir = string.split(os.getenv("PATH"), os.pathsep)
+        wrapexe_exists = False
+        for counter in PATH_dir:
+            if os.path.exists(counter + os.sep + options.wrapexe):
+                wrapexe_exists = True
+                break
+
+        if not wrapexe_exists:
+            error("missing the required launcher: %s \nAdd it to your $PATH\n" % options.wrapexe, 1)
+        
+        if platform.system() == 'Windows' or 'CYGWIN_NT' in platform.system():
+            OS = "Windows"
+        else:
+            if platform.system() == 'Darwin':
+                OS = "Mac"
+            else:
+                OS = "Linux"
+
+        if not (OS  == 'Linux'):
+            error ("knc target supported only on Linux", 1)
     # if no specific test files are specified, run all of the tests in tests/,
     # failing_tests/, and tests_errors/
     if len(args) == 0:
@@ -644,14 +699,21 @@ def run_tests(options1, args, print_version):
     task_threads = [0] * nthreads
     for x in range(nthreads):
         task_threads[x] = multiprocessing.Process(target=run_tasks_from_queue, args=(q, qret, qskip, total_tests,
-            max_test_length, finished_tests_counter, finished_tests_counter_lock, glob_var))
+                max_test_length, finished_tests_counter, finished_tests_counter_lock, glob_var))
         task_threads[x].start()
+
     # wait for them to all finish and then return the number that failed
     # (i.e. return 0 if all is ok)
     for t in task_threads:
         t.join()
     if options.non_interactive == False:
         print_debug("\n", s, run_tests_log)
+
+    
+    for jb in task_threads:
+        if not jb.exitcode == 0:
+            raise OSError(2, 'Some test subprocess has thrown an exception', '')
+
 
     temp_time = (time.time() - start_time)
     elapsed_time = time.strftime('%Hh%Mm%Ssec.', time.gmtime(temp_time))
@@ -724,7 +786,7 @@ if __name__ == "__main__":
     parser.add_option("-f", "--ispc-flags", dest="ispc_flags", help="Additional flags for ispc (-g, -O1, ...)",
                   default="")
     parser.add_option('-t', '--target', dest='target',
-                  help='Set compilation target (sse2-i32x4, sse2-i32x8, sse4-i32x4, sse4-i32x8, sse4-i16x8, sse4-i8x16, avx1-i32x8, avx1-i32x16, avx1.1-i32x8, avx1.1-i32x16, avx2-i32x8, avx2-i32x16, generic-x1, generic-x4, generic-x8, generic-x16, generic-x32, generic-x64)',
+                  help='Set compilation target (sse2-i32x4, sse2-i32x8, sse4-i32x4, sse4-i32x8, sse4-i16x8, sse4-i8x16, avx1-i32x8, avx1-i32x16, avx1.1-i32x8, avx1.1-i32x16, avx2-i32x8, avx2-i32x16, generic-x1, generic-x4, generic-x8, generic-x16, generic-x32, generic-x64, knc)',
                                     default="sse4")
     parser.add_option('-a', '--arch', dest='arch',
                   help='Set architecture (arm, x86, x86-64)',
@@ -749,6 +811,8 @@ if __name__ == "__main__":
                   action = "store_true")
     parser.add_option("--file", dest='in_file', help='file to save run_tests output', default="")
     parser.add_option("--verify", dest='verify', help='verify the file fail_db.txt', default=False, action="store_true")
+    parser.add_option("--save-bin", dest='save_bin', help='compile and create bin, but don\'t execute it',
+                  default=False, action="store_true")
     (options, args) = parser.parse_args()
     L = run_tests(options, args, 1)
     exit(0)
